refactor(banner): move useEffect to Browse
